Refactored tests folder structure.
