+++ src/test/java/com/pika/Onlinegrocerystore/dao/CustomerDaoTests.java
@@ -1,4 +1,4 @@
-package com.pika.Onlinegrocerystore;
+package com.pika.Onlinegrocerystore.dao;
 
 import com.pika.Onlinegrocerystore.dao.ICustomerDao;
 import com.pika.Onlinegrocerystore.domain.Customer;
