+++ src/test/java/com/pika/Onlinegrocerystore/dao/ProductDaoTests.java
@@ -1,4 +1,4 @@
-package com.pika.Onlinegrocerystore;
+package com.pika.Onlinegrocerystore.dao;
 
 import com.pika.Onlinegrocerystore.dao.IProductDao;
 import com.pika.Onlinegrocerystore.domain.Product;
